chore/codygatewayevents: extract into standalone package for reuse, split up internal/codygateway (#63528)
Allows us to directly reuse the Cody Gateway usage queries so that they
can be served directly from Enterprise Portal
(https://github.com/sourcegraph/sourcegraph/pull/63531). To enable this
we also need to split up the monolithic `internal/codygateway` package
so that not all roads lead back to `conf`:

- `internal/codygateway`: Client mechanisms
- `internal/codygateway/codygatewayevents`: Cody Gateway events service
+ related consts
- `internal/codygateway/codygatewayactor`: Cody Gateway actor types

Part of https://linear.app/sourcegraph/issue/CORE-201

## Test plan

n/a
