+++ cmd/cody-gateway/internal/actor/dotcomuser/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//cmd/cody-gateway/internal/limiter",
         "//internal/accesstoken",
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//internal/trace",
         "//lib/errors",
         "@com_github_graph_gophers_graphql_go//:graphql-go",
@@ -35,6 +36,7 @@ go_test(
         "//cmd/cody-gateway/internal/limiter",
         "//internal/accesstoken",
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//lib/errors",
         "@com_github_graph_gophers_graphql_go//relay",
         "@com_github_gregjones_httpcache//:httpcache",
