+++ cmd/cody-gateway/internal/actor/limits_test.go
@@ -11,11 +11,12 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/cmd/cody-gateway/internal/limiter"
 	"github.com/sourcegraph/sourcegraph/internal/codygateway"
+	"github.com/sourcegraph/sourcegraph/internal/codygateway/codygatewayactor"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 func TestNewRateLimitWithPercentageConcurrency(t *testing.T) {
-	concurrencyLimitConfig := codygateway.ActorConcurrencyLimitConfig{
+	concurrencyLimitConfig := codygatewayactor.ActorConcurrencyLimitConfig{
 		Percentage: 0.1,
 		Interval:   10 * time.Second,
 	}
