+++ cmd/cody-gateway/internal/auth/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
         "//cmd/cody-gateway/internal/events",
         "//cmd/cody-gateway/internal/response",
         "//internal/authbearer",
-        "//internal/codygateway",
+        "//internal/codygateway/codygatewayevents",
         "//internal/trace",
         "//lib/errors",
         "@com_github_sourcegraph_log//:log",
@@ -34,7 +34,7 @@ go_test(
         "//cmd/cody-gateway/internal/actor/productsubscription",
         "//cmd/cody-gateway/internal/actor/productsubscription/productsubscriptiontest",
         "//cmd/cody-gateway/internal/events",
-        "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//lib/enterpriseportal/codyaccess/v1:codyaccess",
         "//lib/errors",
         "@com_github_derision_test_go_mockgen_v2//testutil/require",
