+++ cmd/cody-gateway/internal/events/instrumented.go
@@ -33,10 +33,7 @@ func (i *instrumentedLogger) LogEvent(spanCtx context.Context, event Event) erro
 	}
 
 	if err := i.Logger.LogEvent(spanCtx, event); err != nil {
-		if err != nil {
-			span.SetStatus(codes.Error, err.Error())
-		}
-		span.SetStatus(codes.Error, "failed to log event")
+		span.SetStatus(codes.Error, err.Error())
 		return err
 	}
 	return nil
