+++ cmd/cody-gateway/internal/httpapi/attribution/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
         "//cmd/cody-gateway/internal/dotcom",
         "//cmd/cody-gateway/internal/response",
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//internal/trace",
         "//lib/errors",
         "@com_github_khan_genqlient//graphql",
@@ -30,6 +31,7 @@ go_test(
         "//cmd/cody-gateway/internal/events",
         "//cmd/cody-gateway/internal/httpapi",
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "@com_github_google_go_cmp//cmp",
         "@com_github_khan_genqlient//graphql",
         "@com_github_sourcegraph_log//logtest",
