+++ cmd/cody-gateway/internal/httpapi/featurelimiter/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//cmd/cody-gateway/internal/notify",
         "//cmd/cody-gateway/internal/response",
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayevents",
         "//internal/completions/types",
         "//internal/trace",
         "//lib/errors",
