+++ cmd/cody-gateway/internal/notify/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
     visibility = ["//cmd/cody-gateway:__subpackages__"],
     deps = [
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//internal/redislock",
         "//internal/redispool",
         "//internal/trace",
@@ -29,6 +30,7 @@ go_test(
     tags = [TAG_CODY_PRIME],
     deps = [
         "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//internal/redispool",
         "@com_github_hexops_autogold_v2//:autogold",
         "@com_github_slack_go_slack//:slack",
