+++ cmd/cody-gateway/shared/config/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/cody-gateway/internal/httpapi/embeddings",
-        "//internal/codygateway",
+        "//internal/codygateway/codygatewayactor",
         "//internal/collections",
         "//internal/completions/client/anthropic",
         "//internal/completions/client/fireworks",
