+++ cmd/frontend/internal/dotcom/productsubscription/codygateway_graphql.go
@@ -7,7 +7,8 @@ import (
 	"github.com/sourcegraph/log"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
-	"github.com/sourcegraph/sourcegraph/internal/codygateway"
+	"github.com/sourcegraph/sourcegraph/internal/codygateway/codygatewayactor"
+	"github.com/sourcegraph/sourcegraph/internal/codygateway/codygatewayevents"
 	"github.com/sourcegraph/sourcegraph/internal/completions/types"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/gqlutil"
@@ -67,7 +68,7 @@ func (r codyGatewayAccessResolver) ChatCompletionsRateLimit(ctx context.Context)
 	return &codyGatewayRateLimitResolver{
 		feature:     types.CompletionsFeatureChat,
 		actorID:     r.sub.UUID(),
-		actorSource: codygateway.ActorSourceEnterpriseSubscription,
+		actorSource: codygatewayactor.ActorSourceEnterpriseSubscription,
 		v:           rateLimit,
 		source:      source,
 	}, nil
@@ -107,7 +108,7 @@ func (r codyGatewayAccessResolver) CodeCompletionsRateLimit(ctx context.Context)
 	return &codyGatewayRateLimitResolver{
 		feature:     types.CompletionsFeatureCode,
 		actorID:     r.sub.UUID(),
-		actorSource: codygateway.ActorSourceEnterpriseSubscription,
+		actorSource: codygatewayactor.ActorSourceEnterpriseSubscription,
 		v:           rateLimit,
 		source:      source,
 	}, nil
@@ -146,7 +147,7 @@ func (r codyGatewayAccessResolver) EmbeddingsRateLimit(ctx context.Context) (gra
 
 	return &codyGatewayRateLimitResolver{
 		actorID:     r.sub.UUID(),
-		actorSource: codygateway.ActorSourceEnterpriseSubscription,
+		actorSource: codygatewayactor.ActorSourceEnterpriseSubscription,
 		v:           rateLimit,
 		source:      source,
 	}, nil
@@ -154,7 +155,7 @@ func (r codyGatewayAccessResolver) EmbeddingsRateLimit(ctx context.Context) (gra
 
 type codyGatewayRateLimitResolver struct {
 	actorID     string
-	actorSource codygateway.ActorSource
+	actorSource codygatewayactor.ActorSource
 	feature     types.CompletionsFeature
 	source      graphqlbackend.CodyGatewayRateLimitSource
 	v           licensing.CodyGatewayRateLimit
@@ -174,7 +175,7 @@ func (r *codyGatewayRateLimitResolver) IntervalSeconds() int32 { return r.v.Inte
 
 func (r codyGatewayRateLimitResolver) Usage(ctx context.Context) ([]graphqlbackend.CodyGatewayUsageDatapoint, error) {
 	var (
-		usage []SubscriptionUsage
+		usage []codygatewayevents.SubscriptionUsage
 		err   error
 	)
 	if r.feature != "" {
