+++ internal/codygateway/codygatewayactor/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "codygatewayactor",
+    srcs = ["actor.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/internal/codygateway/codygatewayactor",
+    visibility = ["//:__subpackages__"],
+)
